+++ docker-compose.yaml
@@ -22,7 +22,7 @@ services:
     environment:
       PORT: 3000
       NODE_ENV: development
-      ENDPOINT_URL: http://localstack:4566
+      SQS_ENDPOINT: http://localstack:4566
       MONGO_URI: mongodb://mongodb:27017/
       DATA_INGEST_QUEUE_URL: http://sqs.eu-west-2.127.0.0.1:4566/000000000000/fcp_sfd_data_ingest.fifo
       DATA_INGEST_DEAD_LETTER_QUEUE_URL: http://sqs.eu-west-2.127.0.0.1:4566/000000000000/fcp_sfd_data_ingest-deadletter.fifo
