+++ src/config/index.js
@@ -1,199 +1,15 @@
 import convict from 'convict'
-import path from 'node:path'
-import { fileURLToPath } from 'node:url'
-import environments from '../constants/environments.js'
 
-const dirname = path.dirname(fileURLToPath(import.meta.url))
-
-const isProduction = process.env.NODE_ENV === environments.PRODUCTION
-const isDev = process.env.NODE_ENV === environments.DEVELOPMENT
-const isTest = process.env.NODE_ENV === environments.TEST
+import { serverConfig } from './server.js'
+import { awsConfig } from './aws.js'
+import { mongoConfig } from './mongo.js'
+import { messagingConfig } from './messaging.js'
 
 const config = convict({
-  serviceVersion: {
-    doc: 'The service version, this variable is injected into your docker container in CDP environments',
-    format: String,
-    nullable: true,
-    default: null,
-    env: 'SERVICE_VERSION'
-  },
-  env: {
-    doc: 'The application environment.',
-    format: ['production', 'development', 'test'],
-    default: 'development',
-    env: 'NODE_ENV'
-  },
-  port: {
-    doc: 'The port to bind.',
-    format: 'port',
-    default: 3555,
-    env: 'PORT'
-  },
-  serviceName: {
-    doc: 'Api Service Name',
-    format: String,
-    default: 'fcp-sfd-data'
-  },
-  root: {
-    doc: 'Project root',
-    format: String,
-    default: path.resolve(dirname, '../..')
-  },
-  isProduction: {
-    doc: 'If this application running in the production environment',
-    format: Boolean,
-    default: isProduction
-  },
-  isDevelopment: {
-    doc: 'If this application running in the development environment',
-    format: Boolean,
-    default: isDev
-  },
-  isTest: {
-    doc: 'If this application running in the test environment',
-    format: Boolean,
-    default: isTest
-  },
-  log: {
-    enabled: {
-      doc: 'Is logging enabled',
-      format: Boolean,
-      default: !isTest,
-      env: 'LOG_ENABLED'
-    },
-    level: {
-      doc: 'Logging level',
-      format: ['fatal', 'error', 'warn', 'info', 'debug', 'trace', 'silent'],
-      default: 'info',
-      env: 'LOG_LEVEL'
-    },
-    format: {
-      doc: 'Format to output logs in.',
-      format: ['ecs', 'pino-pretty'],
-      default: isProduction ? 'ecs' : 'pino-pretty',
-      env: 'LOG_FORMAT'
-    },
-    redact: {
-      doc: 'Log paths to redact',
-      format: Array,
-      default: isProduction
-        ? ['req.headers.authorization', 'req.headers.cookie', 'res.headers']
-        : ['req', 'res', 'responseTime']
-    }
-  },
-  mongoUri: {
-    doc: 'URI for mongodb',
-    format: String,
-    default: 'mongodb://127.0.0.1:27017/',
-    env: 'MONGO_URI'
-  },
-  mongoDatabase: {
-    doc: 'database for mongodb',
-    format: String,
-    default: 'fcp-sfd-data',
-    env: 'MONGO_DATABASE'
-  },
-  httpProxy: {
-    doc: 'HTTP Proxy',
-    format: String,
-    nullable: true,
-    default: null,
-    env: 'HTTP_PROXY'
-  },
-  isSecureContextEnabled: {
-    doc: 'Enable Secure Context',
-    format: Boolean,
-    default: isProduction,
-    env: 'ENABLE_SECURE_CONTEXT'
-  },
-  isMetricsEnabled: {
-    doc: 'Enable metrics reporting',
-    format: Boolean,
-    default: isProduction,
-    env: 'ENABLE_METRICS'
-  },
-  tracing: {
-    header: {
-      doc: 'Which header to track',
-      format: String,
-      default: 'x-cdp-request-id',
-      env: 'TRACING_HEADER'
-    }
-  },
-  aws: {
-    endpoint: {
-      doc: 'AWS Endpoint (for LocalStack)',
-      format: String,
-      default: null,
-      env: 'ENDPOINT_URL'
-    },
-    region: {
-      doc: 'AWS Region',
-      format: String,
-      default: 'eu-west-2',
-      env: 'AWS_REGION'
-    },
-    accessKeyId: {
-      doc: 'AWS Access Key ID',
-      format: String,
-      default: null,
-      nullable: true,
-      env: 'AWS_ACCESS_KEY_ID'
-    },
-    secretAccessKey: {
-      doc: 'AWS Secret Access Key',
-      format: String,
-      default: null,
-      nullable: true,
-      env: 'AWS_SECRET_ACCESS_KEY'
-    }
-  },
-  messaging: {
-    waitTimeSeconds: {
-      doc: 'SQS Consumer for which to wait for messages',
-      format: Number,
-      default: 10,
-      env: 'SQS_CONSUMER_WAIT_TIME_SECONDS'
-    },
-    batchSize: {
-      doc: 'The maximum number of messages to return in each call',
-      format: Number,
-      default: 10,
-      env: 'SQS_CONSUMER_BATCH_SIZE'
-    },
-    pollingWaitTime: {
-      doc: 'The duration (in seconds) before sqs-consumer polls for new messages',
-      format: Number,
-      default: 0,
-      env: 'SQS_CONSUMER_POLLING_WAIT_TIME'
-    },
-    visibilityTimeout: {
-      doc: 'The duration (in seconds) that the received messages are hidden from subsequent retrieve requests after being retrieved by a ReceiveMessage request',
-      format: Number,
-      default: 30,
-      env: 'SQS_CONSUMER_VISIBILITY_TIMEOUT'
-    },
-    heartbeatInterval: {
-      doc: 'The interval (in seconds) between requests to extend the message visibility timeout. Must be less than the visibility timeout',
-      format: Number,
-      default: 5,
-      env: 'SQS_CONSUMER_HEARTBEAT_INTERVAL'
-    },
-    dataIngestion: {
-      queueUrl: {
-        doc: 'URL for the data ingest queue',
-        format: String,
-        default: null,
-        env: 'DATA_INGEST_QUEUE_URL'
-      },
-      deadLetterUrl: {
-        doc: 'URL for the data ingest dead letter queue',
-        format: String,
-        default: null,
-        env: 'DATA_INGEST_DEAD_LETTER_QUEUE_URL'
-      }
-    }
-  }
+  ...serverConfig,
+  ...awsConfig,
+  ...mongoConfig,
+  ...messagingConfig
 })
 
 config.validate({ allowed: 'strict' })
